# src/main/resources/swagger/api.yml
swagger: "2.0"
info:
  description: "This is a specification for VehicleLogistics API"
  version: "1.0"
  title: "Vehicle Logistics Application"
  contact:
    email: "arvaroh73@gmail.com"
host: "localhost:8080"
basePath: "/vehicle"
tags:
  - name: "vehicle-controller"
    description: "Everything about Vehicle in the system"
schemes:
  - "http"
paths:
  /location:
    post:
      tags:
        - "vehicle-controller"
      summary: "Add a new vehicle with location"
      description: ""
      operationId: "createCarLocation"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: "body"
          name: "VehicleDto"
          description: "Vehicle object that needs to be added to DB"
          required: true
          schema:
            $ref: "#/definitions/VehicleDto"
      responses:
        200:
          description: "Location is changed successfully"
        400:
          description: "Invalid input"
  /vehicles:
    get:
      tags:
        - "vehicle-controller"
      summary: "Get list of vehicles within rectangle"
      description: ""
      operationId: "getVehicleList"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - in: query
          name: "bottomLatitude"
          description: "bottom latitude"
          required: true
          type: "number"
          format: "double"
          minimum: -90.00
          maximum: 90.00
        - in: query
          name: "topLatitude"
          description: "top latitude"
          required: true
          type: "number"
          format: "double"
          minimum: -90.00
          maximum: 90.00
        - in: query
          name: "leftLongitude"
          description: "left longitude"
          required: true
          type: "number"
          format: "double"
          minimum: -180.00
#          exclusiveMinimum: true
          maximum: 180.00
        - in: query
          name: "rightLongitude"
          description: "right longitude"
          required: true
          type: "number"
          format: "double"
          minimum: -180.00
#          exclusiveMinimum: true
          maximum: 180.00
      responses:
        200:
          description: OK
          schema:
            type: array
            items:
              type: "integer"
              format: "int32"
#            $ref: '#/definitions/VehicleListDto'
        400:
          description: "Invalid input"

definitions:
  VehicleDto:
    type: "object"
    properties:
      vehicleId:
        type: "integer"
        format: "int32"
        minimum: 1
      latitude:
        type: "number"
        format: "double"
        minimum: -90.00
        maximum: 90.00
      longitude:
        type: "number"
        format: "double"
        minimum: -180.00
        maximum: 180.00
    xml:
      name: "VehicleDto"
#  VehicleListDto:
#    type: "object"
#    properties:
#      appointmentSlotDtoList:
#        type: array
#        items:
#          type: "integer"
#          format: "int32"
#  Coordinate:
#    type: "object"
#    properties:
#      latitude:
#        type: "number"
#        format: "float"
#        minimum: -90
#        maximum: 90
#      longitude:
#        type: "number"
#        format: "float"
#        minimum: -180
#        maximum: 180
#    xml:
#      name: "Coordinate"